requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg[binary]==3.2.9
celery==5.4.0
redis==5.0.7
python-dotenv==1.0.1
dj-database-url==2.2.0
gunicorn==22.0.0
uvicorn[standard]==0.30.1